feat: update products folder
