+++ src/schema/resolvers/Product/Query/index.ts
@@ -1,0 +1,5 @@
+import products from './products'
+
+export default {
+  products,
+}
